Add files via upload
The following small simulation shows that all forms of centering (mean centering and residual centering) do not have an effect of the main target of interaction modeling—that is the B coefficient of the product term as well as its standard errors and t-value. The coefficients of the first-order effects, of course, change in respect to the centering procedure but that is not a consequence of bias. Instead, centering effects the meaning of the first order effects in presence of a product term